Fixed taking the correct interval
The problem was with taking the max of the intervals, which were tuples of (x, y) values. We have to take the biggest starting interval, therefore biggest x, and then the smallest ending interval, meaning y. An example which did not work before and which put all comments in the beginning of the code after unparse is performed is:

code = """
if 1 == 1:
   # Test1
   print('a')
else:
   # Test2
   print('b')
print('b')
"""